+++ src/EVM/FeeSchedule.hs
@@ -1,5 +1,7 @@
 module EVM.FeeSchedule where
 
+import Data.Word (Word64)
+
 data FeeSchedule n = FeeSchedule
   { g_zero :: n
   , g_base :: n
@@ -51,6 +53,7 @@ data FeeSchedule n = FeeSchedule
   , g_warm_storage_read :: n
   , g_access_list_address :: n
   , g_access_list_storage_key :: n
+  , g_txdatafloor :: n
   } deriving Show
 
 feeSchedule :: Num n => FeeSchedule n
@@ -105,4 +108,8 @@ feeSchedule = FeeSchedule
   , g_warm_storage_read = 100
   , g_access_list_address = 2400
   , g_access_list_storage_key = 1900
+  , g_txdatafloor = 10
   }
+
+defaultTxdataFloorGas :: Word64
+defaultTxdataFloorGas = feeSchedule.g_transaction