+++ src/EVM/Types.hs
@@ -270,6 +270,7 @@ data Expr (a :: EType) where
   SHL            :: Expr EWord -> Expr EWord -> Expr EWord
   SHR            :: Expr EWord -> Expr EWord -> Expr EWord
   SAR            :: Expr EWord -> Expr EWord -> Expr EWord
+  CLZ            :: Expr EWord -> Expr EWord
 
   -- Hashes
 
@@ -836,6 +837,7 @@ data TxState = TxState
   , subState    :: SubState
   , isCreate    :: Bool
   , txReversion :: Map (Expr EAddr) Contract
+  , txdataFloorGas :: Word64
   }
   deriving (Show)
 
@@ -1019,6 +1021,8 @@ data VMOpts (t :: VMType) = VMOpts
   , allowFFI :: Bool
   , freshAddresses :: Int
   , beaconRoot :: W256
+  , parentHash :: W256
+  , txdataFloorGas :: Word64
   }
 
 deriving instance Show (VMOpts Symbolic)
@@ -1057,6 +1061,7 @@ data GenericOp a
   | OpShl
   | OpShr
   | OpSar
+  | OpClz
   | OpSha3
   | OpAddress
   | OpBalance